+++ tests/laravel/app/Traits/Searchable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Traits;
+
+use Laravel\Scout\Searchable as SearchableParent;
+
+trait Searchable
+{
+    use SearchableParent;
+}